pyproject.toml
[tool.poetry]
name = "prefect_docker"
version = "3.0.1"
description = "Workflow for metadata ingestion into Dataverse"
authors = ["Fjodor van Rijsselberg"]

[tool.poetry.dependencies]
python = "^3.12"
prefect = "^3.0.0"
python-dotenv = "^1.0.0"
boto3 = "^1.35.00"
dynaconf = "^3.2.0"
jmespath = "^1.0.1"
pydantic = "^2.9.0"
pydataverse = "^0.3.4"
openpyxl = "^3.1.5"
asyncpg = "^0.29.0"
pillow = "^11.0.0"
numpy = "^2.2.1"
keybert = "^0.8.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
